+++ CHANGELOG.md
@@ -30,6 +30,8 @@
 
 ### Features added:
 * Adjust UI for the Field protection create/edit screen (UIDATIMP-1043)
+* Cover `<MatchProfiles>` settings with tests (UIDATIMP-975)
+* Cover `<ProfileAssociator>` component with tests (UIDATIMP-717)
 
 ## [5.0.1](https://github.com/folio-org/ui-data-import/tree/v5.0.1) (2021-10-19)